+++ frontend/src/index.tsx
@@ -4,15 +4,19 @@ import { App } from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from "./redux/store";
+import { store } from './redux/store';
+import { ThemeProvider } from '@mui/material/styles';
+import { theme } from './theme';
 
 ReactDOM.render(
-  <BrowserRouter>
-      <Provider store={store}>
-    <App />
-      </Provider>
-  </BrowserRouter>,
-  document.getElementById('root')
+	<BrowserRouter>
+		<ThemeProvider theme={theme}>
+			<Provider store={store}>
+				<App/>
+			</Provider>
+		</ThemeProvider>
+	</BrowserRouter>,
+	document.getElementById('root')
 );
 
 reportWebVitals();