+++ frontend/src/pages/Profile/Profile.tsx
@@ -4,13 +4,13 @@ import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../redux/store";
 import { fetchedProfile } from "../../redux/ducks/profile/profile.slice";
 import {  useNavigate, useParams, useSearchParams } from "react-router-dom";
-import { Card, Grid } from "@mui/material";
+import { Grid } from "@mui/material";
 // @ts-ignore
 import styles from "./Profile.module.css";
 import { ProfileInfo } from "./ProfileInfo";
 import { ProfileHeader } from "./ProfileHeader/ProfileHeader";
 import { ProfileEntities } from "./ProfileEntities";
-import { ProfileGallery } from "./ProfileGallery";
+import { ProfileGallery } from "./ProfileGallery/ProfileGallery";
 
 interface ProfileInterface {
 	isAuth: boolean;
@@ -19,7 +19,7 @@ interface ProfileInterface {
 
 export const Profile: React.FC<ProfileInterface> = ({ isAuth, userId }) => {
 	let navigate = useNavigate();
-	const { profile } = useSelector((state: RootState) => state.profileReducer);
+	const { profile, error, loading } = useSelector((state: RootState) => state.profileReducer);
 	let url = useParams<string>();
 	const [searchParams, setSearchParams] = useSearchParams();
 	const params = searchParams.get('gallery')
@@ -43,13 +43,13 @@ export const Profile: React.FC<ProfileInterface> = ({ isAuth, userId }) => {
 
 
 	return (
-		<Card className={styles.root}>
+		<Grid className={styles.root}>
 			{gallery && <ProfileGallery gallery={gallery} setGallery={setGallery} userId={profile.id} images={profile.photos}/>}
 			<Grid className={styles.profileHeader}>
 				<ProfileHeader isOwner={isOwner} profile={profile}/>
 			</Grid>
 			<Grid className={styles.profileInfo}>
-				<ProfileInfo profile={profile} />
+				<ProfileInfo profile={profile} error={error} />
 			</Grid>
 			<Grid className={styles.profilePosts}>
 				<ProfilePosts isOwner={isOwner} userId={userId} posts={profile?.posts} />
@@ -58,6 +58,6 @@ export const Profile: React.FC<ProfileInterface> = ({ isAuth, userId }) => {
 				<ProfileEntities profile={profile} />
 			</Grid>
 
-		</Card>
+		</Grid>
 	);
 };